Bug 338 - Have ddlScript return a bigint since the action_seq is a bigint

+++ src/backend/slony1_funcs.sql
@@ -3615,7 +3615,7 @@ Set sequence seq_id to have new value last_value.
 --	Capture DDL into sl_log_script
 -- ----------------------------------------------------------------------
 create or replace function @NAMESPACE@.ddlCapture (p_statement text, p_nodes text)
-returns integer
+returns bigint
 as $$
 declare
 	c_local_node	integer;
